// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
      admin_users: {
        Row: {
          created_at: string
          id: string
          role: string
        }
        Insert: {
          created_at?: string
          id: string
          role?: string
        }
        Update: {
          created_at?: string
          id?: string
          role?: string
        }
        Relationships: []
      }
      alerts: {
        Row: {
          alert_type: Database["public"]["Enums"]["alert_type"]
          created_at: string
          created_by: string | null
          description: string
          end_time: string | null
          id: string
          latitude: number
          longitude: number
          radius: number
          severity: Database["public"]["Enums"]["alert_severity"]
          source: string | null
          start_time: string
          title: string
          updated_at: string
        }
        Insert: {
          alert_type: Database["public"]["Enums"]["alert_type"]
          created_at?: string
          created_by?: string | null
          description: string
          end_time?: string | null
          id?: string
          latitude: number
          longitude: number
          radius: number
          severity: Database["public"]["Enums"]["alert_severity"]
          source?: string | null
          start_time?: string
          title: string
          updated_at?: string
        }
        Update: {
          alert_type?: Database["public"]["Enums"]["alert_type"]
          created_at?: string
          created_by?: string | null
          description?: string
          end_time?: string | null
          id?: string
          latitude?: number
          longitude?: number
          radius?: number
          severity?: Database["public"]["Enums"]["alert_severity"]
          source?: string | null
          start_time?: string
          title?: string
          updated_at?: string
        }
        Relationships: []
      }
// ...
    }
// ...
    Enums: {
      alert_severity: "low" | "medium" | "high" | "critical"
      alert_type: "police" | "fire" | "health" | "weather" | "other"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never